@@ -5,7 +5,13 @@ module.exports = {
         ['@babel/preset-env', { 'targets': { 'node': 'current' }}]
       ],
       'plugins': [
-        '@babel/plugin-syntax-dynamic-import'
+        '@babel/plugin-syntax-dynamic-import',
+        [
+          '@babel/plugin-proposal-class-properties',
+          {
+            'loose': true
+          }
+        ]
       ]
     }
   }
