@@ -1,10 +1,13 @@
 const { path } = require('@vuepress/shared-utils')
 
-module.exports = (options, context) => ({
+/**
+ * @type {import('@vuepress/types').Plugin<{}, import('@vuepress/types').DefaultThemeConfig>}
+ */
+module.exports = (options, ctx) => ({
   name: '@vuepress/internal-enhance-app',
 
   enhanceAppFiles () {
-    const { sourceDir, themeAPI } = context
+    const { sourceDir, themeAPI } = ctx
     const enhanceAppPath = path.resolve(sourceDir, '.vuepress/enhanceApp.js')
     const files = [enhanceAppPath]
     if (themeAPI.existsParentTheme) {
