@@ -1,11 +1,17 @@
 const { logger, fs, path } = require('@vuepress/shared-utils')
 
-module.exports = ({
-  serviceWorker = true,
-  updatePopup = false,
-  popupComponent = 'SWUpdatePopup',
-  generateSWConfig = {}
-}, context) => ({
+/**
+ * @type {import('@vuepress/types').Plugin}
+ */
+module.exports = (
+  {
+    serviceWorker = true,
+    updatePopup = false,
+    popupComponent = 'SWUpdatePopup',
+    generateSWConfig = {}
+  },
+  context
+) => ({
   alias: {
     '@sw-event': path.resolve(__dirname, 'lib/event.js')
   },
@@ -33,12 +39,17 @@ module.exports = ({
       await wbb.generateSW({
         swDest: swFilePath,
         globDirectory: outDir,
-        globPatterns: ['**\/*.{js,css,html,png,jpg,jpeg,gif,svg,woff,woff2,eot,ttf,otf}'],
+        globPatterns: [
+          '**/*.{js,css,html,png,jpg,jpeg,gif,svg,woff,woff2,eot,ttf,otf}'
+        ],
         ...generateSWConfig
       })
       await fs.writeFile(
         swFilePath,
-        await fs.readFile(path.resolve(__dirname, 'lib/skip-waiting.js'), 'utf8'),
+        await fs.readFile(
+          path.resolve(__dirname, 'lib/skip-waiting.js'),
+          'utf8'
+        ),
         { flag: 'a' }
       )
     }
