@@ -53,9 +53,10 @@ export default {
           {
             inputSelector: '#algolia-search-input',
             // #697 Make docsearch work well at i18n mode.
-            algoliaOptions: Object.assign({
-              'facetFilters': [`lang:${lang}`].concat(algoliaOptions.facetFilters || [])
-            }, algoliaOptions),
+            algoliaOptions: {
+              ...algoliaOptions,
+              facetFilters: [`lang:${lang}`].concat(algoliaOptions.facetFilters || [])
+            },
             handleSelected: (input, event, suggestion) => {
               const { pathname, hash } = new URL(suggestion.url)
               const routepath = pathname.replace(this.$site.base, '/')
