@@ -1,23 +1,28 @@
 module.exports = {
   rules: {
-    '@textlint-rule/no-unmatched-pair': true,
+    "@textlint-rule/no-unmatched-pair": true,
     apostrophe: true,
-    'common-misspellings': true,
+    "common-misspellings": true,
     diacritics: true,
-    'en-capitalization': {
-      allowHeading: false
-    },
-    'stop-words': {
-      severity: 'warning'
+    "en-capitalization": false,
+    "stop-words": {
+      severity: "warning"
     },
     terminology: {
-      terms: `${__dirname}/.textlint.terms.json`
+      terms: [
+        "VuePress",
+        "sass",
+        [
+          "front[- ]matter",
+          "frontmatter"
+        ]
+      ]
     },
-    'write-good': {
-      severity: 'warning'
+    "write-good": {
+      severity: "warning"
     }
   },
   filters: {
     comments: true
   }
-}
+};
